+++ apiEx.js
@@ -1,6 +1,6 @@
 const ApiWrapper = require('tradelayer');
-let myInfo = {address:'ltc1qehzkx0fpdydj48njs63hyqu02luzcxn66rtqjj',otherAddrs:[]};
-const api = new ApiWrapper('ws://172.81.181.19', 3001, false,true, myInfo);
+let myInfo = {address:'tltc1qh4se4w23draju8ef82vdvelz3zj8egflrg2gve',otherAddrs:[]};
+const api = new ApiWrapper('ws://172.81.181.19', 3001, false,true, myInfo, 'LTCTEST');
 
 // Start listening for order matches and handle swaps
 let orderbookSession = []
