+++ package.json
@@ -1,32 +1,21 @@
 {
   "name": "tradelayer",
-  "version": "1.4.0",
+  "version": "1.5.1",
   "description": "NPM package for interacting with TradeLayer orderbook and building Litecoin/Token/Futures transactions.",
   "main": "index.js",
+  {
   "dependencies": {
-    "abstract-level": "^1.0.3",
     "axios": "^1.7.7",
-    "big-json": "^3.2.0",
-    "bigint-base-converter": "0.1.3",
-    "bignumber": "^1.1.0",
     "bignumber.js": "^9.1.2",
-    "bitcoin": "3.0.3",
-    "bitcoinjs-lib": "*",
+    "bitcoinjs-lib": "^6.1.6",
     "bitcore-lib-ltc": "^10.0.21",
-    "buffer": "^6.0.3",
-    "child_process": "^1.0.2",
-    "dogecoind-rpc": "0.8.1",
-    "dotenv": "^16.0.0",
-    "ecpair": "*",
-    "express": "^4.18.2",
-    "litecoin": "^2.0.5",
-    "nedb": "^1.8.0",
-    "node-fetch": "^2.7.0",
-    "socket.io-client": "^4.1.3",
-    "tiny-secp256k1": "^2.2.3",
-    "util": "^0.12.4",
-    "uuid": "^9.0.1"
-  },
+    "dotenv": "^16.4.5",
+    "isomorphic-ws": "^5.0.0",
+    "ws": "^8.18.0",
+    "ccxt": "^4.4.77",
+    "socket.io-client": "^4.7.5"
+  }
+},
   "repository": {
     "type": "git",
     "url": "https://github.com/patrickdugan/tlnpm"
